+++ tsconfig.json
@@ -39,8 +39,8 @@
         "./dist/modules/common/interceptors/*.interceptor"
       ],
       "@services/*": [
-        "./src/modules/common/services/*.service",
-        "./dist/modules/common/services/*.service"
+        "./src/services/*.service",
+        "./dist/services/*.service"
       ]
     }
   },
